# docs/openapi.yaml
openapi: 3.1.0
info:
  title: VendorIQ API
  version: 0.1.0
  summary: Vendor management & market intelligence for construction tendering.
  description: |
    The public REST contract of VendorIQ. **Every UI action goes through this API** — the
    frontend holds no business logic (brief §2), so anything a screen can do is an operation
    below, and any other product can do the same with an API key.

    This document is hand-written and is the source of truth: `apps/api` serves it verbatim at
    `/api/openapi.json`, `/api/openapi.yaml` and `/api/docs`, and route handlers are validated
    against it rather than generating it. Changing it is a contract change — see
    `CONTRIBUTING.md`.

    ### Conventions

    * **Ids** are UUIDs, stable across systems. `external_ref` on vendors and projects maps to
      a foreign system's own id.
    * **Errors** always use one envelope: `{"error": {"code", "message", "details"}}`.
    * **Collections** always return `{"items", "total", "page", "page_size"}`.
    * **Money** is AZN. The Rev4 methodology sheet labels turnover USD; the data is AZN and the
      system stores AZN (spec §16).
    * **Language** — user-facing text carries `_az` and `_en` fields; the server never returns a
      single pre-translated sentence for the UI, it returns keys or both languages.
    * **Field codes** (`A.1` … `G.7`, tables `C.t1`, `C.t2`, `G.t1`) come from spec Appendix A;
...
      model version.
...
  license:
    name: Proprietary
  contact:
...
servers:
...
security:
  - sessionCookie: []
  - apiKey: []
paths:
...
  /integrations/adapters/{adapter}/vendors/{vendor_id}/config:
    parameters:
      - { $ref: "#/components/parameters/AdapterKey" }
      - { $ref: "#/components/parameters/VendorId" }
    get:
      tags: [integrations]
      operationId: getAdapterConfig
      summary: Per-vendor connector configuration
      description: Secrets are returned masked; a masked value sent back unchanged leaves the stored secret alone.
      responses:
        "200":
          description: Configuration.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdapterConfig" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [integrations]
      operationId: putAdapterConfig
      summary: Configure the connector for one vendor
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdapterConfigInput" }
      responses:
        "200":
          description: Stored.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdapterConfig" }
  /integrations/adapters/{adapter}/sync:
    parameters: [{ $ref: "#/components/parameters/AdapterKey" }]
    post:
      tags: [integrations]
      operationId: runSync
      summary: Run an adapter now
      description: Without `vendor_id` the adapter runs for every vendor it is configured for.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                vendor_id: { type: string, format: uuid }
                since: { type: string, format: date-time }
      responses:
        "202":
          description: The sync log entry for this run.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SyncLog" }
        "409":
          description: The adapter is not configured.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /integrations/sync-log:
    get:
      tags: [integrations]
      operationId: listSyncLog
      summary: Adapter run history
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: adapter
          in: query
          schema: { type: string }
        - name: vendor_id
          in: query
          schema: { type: string, format: uuid }
        - name: result
          in: query
          schema: { $ref: "#/components/schemas/SyncResult" }
      responses:
        "200":
          description: Runs, newest first.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SyncLogPage" }
  /integrations/excel-import/preview:
    post:
      tags: [integrations]
      operationId: previewExcelImport
      summary: Parse an uploaded workbook and show the mapping
      description: |
        Step one of the two-step import: nothing is written. The response carries the mapped
        fields, the document statuses read from the checklist sheet, the derived raw indicators
        and the anomaly warnings (brief §1.11). The officer confirms, then calls
        `createExcelImportRun` with the returned `preview_id`.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: .xlsx application form or scoring workbook.
                kind:
                  type: string
                  enum: [application_form, scoring_workbook]
                  default: application_form
                vendor_id:
                  type: string
                  format: uuid
                  description: Target vendor; omit to let the parser match on VÖEN.
      responses:
        "200":
          description: The mapping preview.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportPreview" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /integrations/excel-import/runs:
    post:
      tags: [integrations]
      operationId: createExcelImportRun
      summary: Write a confirmed preview into the register
      description: Writes field observations with source `excel` and returns the resulting sync log entry.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [preview_id]
              properties:
                preview_id: { type: string, format: uuid }
                vendor_id:
                  type: string
                  format: uuid
                  description: Overrides the vendor the preview matched.
                accept_field_codes:
                  type: array
                  items: { type: string }
                  description: Import only these codes; omit to import everything in the preview.
      responses:
        "201":
          description: Imported.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SyncLog" }
        "404": { $ref: "#/components/responses/NotFound" }
        # The preview exists but can no longer be acted on: already imported, or past its
        # expiry. `details.reason` is `consumed` or `expired`.
        "409": { $ref: "#/components/responses/Conflict" }
  /integrations/api-keys:
    get:
      tags: [integrations]
      operationId: listApiKeys
      summary: API keys
      description: The key material itself is never returned again after creation.
      responses:
        "200":
          description: Keys.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ApiKey" }
    post:
      tags: [integrations]
      operationId: createApiKey
      summary: Create an API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, scopes]
              properties:
                name: { type: string }
                scopes:
                  type: array
                  items: { $ref: "#/components/schemas/Scope" }
      responses:
        "201":
          description: Created — `key` is shown this once only.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiKeyCreated" }
  /integrations/api-keys/{api_key_id}:
    parameters:
      - name: api_key_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [integrations]
      operationId: patchApiKey
      summary: Rename a key or change its scopes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                scopes:
                  type: array
                  items: { $ref: "#/components/schemas/Scope" }
                is_active: { type: boolean }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiKey" }
    delete:
      tags: [integrations]
      operationId: revokeApiKey
      summary: Revoke a key
      responses:
        "204": { description: Revoked. }
  /integrations/webhooks:
    get:
      tags: [integrations]
      operationId: listWebhooks
      summary: Webhook subscriptions
      responses:
        "200":
          description: Subscriptions.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Webhook" }
    post:
      tags: [integrations]
      operationId: createWebhook
      summary: Subscribe to domain events
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WebhookInput" }
      responses:
        "201":
          description: Created — `secret` is shown this once only.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WebhookCreated" }
  /integrations/webhooks/{webhook_id}:
    parameters:
      - name: webhook_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [integrations]
      operationId: patchWebhook
      summary: Update a subscription
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WebhookInput" }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Webhook" }
    delete:
      tags: [integrations]
      operationId: deleteWebhook
      summary: Remove a subscription
      responses:
        "204": { description: Deleted. }
  /integrations/webhooks/{webhook_id}/test:
    parameters:
      - name: webhook_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [integrations]
      operationId: testWebhook
      summary: Send a signed test delivery
      responses:
        "200":
          description: What the endpoint answered.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WebhookDelivery" }

  # ────────────────────────────────── admin ──────────────────────────────────
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: vendoriq_session
      description: httpOnly session cookie issued by the auth endpoints. Mutating requests also require the `X-CSRF-Token` header.
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Machine credential with scopes; see `/integrations/api-keys`.
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSizeParam:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 25 }
...
    VendorId:
      name: vendor_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    AdapterKey:
      name: adapter
      in: path
      required: true
      schema: { $ref: "#/components/schemas/AdapterKey" }
  responses:
...
    NotFound:
      description: No such resource, or it is outside the caller's scope.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    ValidationError:
      description: The payload failed validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    UnsupportedMediaType:
      description: Only PDF documents and .xlsx workbooks are accepted.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
...
    SyncResult:
      type: string
      enum: [success, partial, failed]
...
    AdapterConfig:
      type: object
      required: [adapter, vendor_id, is_enabled]
      properties:
        adapter: { $ref: "#/components/schemas/AdapterKey" }
        vendor_id: { type: string, format: uuid }
        is_enabled: { type: boolean }
        base_url: { type: [string, "null"], format: uri }
        auth_type: { type: string, enum: [none, basic, bearer, api_key] }
        username: { type: [string, "null"] }
        secret_masked: { type: [string, "null"], description: Never the real secret. }
        field_map:
          type: object
          additionalProperties: { type: string }
          description: Remote path → VendorIQ field code.
        schedule_cron: { type: [string, "null"] }
    AdapterConfigInput:
      type: object
      properties:
        is_enabled: { type: boolean }
        base_url: { type: string, format: uri }
        auth_type: { type: string, enum: [none, basic, bearer, api_key] }
        username: { type: string }
        secret: { type: string, description: Send the masked value back unchanged to keep the stored secret. }
        field_map:
          type: object
          additionalProperties: { type: string }
        schedule_cron: { type: string }
    SyncLog:
      type: object
      required: [id, adapter, started_at, fields_written, result]
      properties:
        id: { type: string, format: uuid }
        adapter: { $ref: "#/components/schemas/AdapterKey" }
        vendor_id: { type: [string, "null"], format: uuid }
        vendor_name: { type: [string, "null"] }
        started_at: { type: string, format: date-time }
        finished_at: { type: [string, "null"], format: date-time }
        fields_written: { type: integer }
        warnings:
          type: array
          items: { $ref: "#/components/schemas/SyncWarning" }
        result: { $ref: "#/components/schemas/SyncResult" }

    # What an adapter *run* can report. A superset of `ImportWarning`, not a sibling: the
    # Excel importer is an adapter like any other, so a run's log carries parser anomalies
    # (`stale_certificate`, addressed to a sheet and a cell) as readily as transport failures
    # (`source_unreachable`, addressed to nothing of the sort). The location fields are
    # therefore optional here and required nowhere.
    #
    # `ImportPreview.warnings` keeps the narrower `ImportWarning`, which is the useful half of
    # the distinction: parsing a workbook cannot fail to reach a host, so a client rendering a
    # preview will never meet the transport codes.
...
    SyncLogPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/SyncLog" }
...